+++ src/context/Master/Mail/Domain/RepositoryMain.ts
@@ -3,5 +3,5 @@ import { Context } from 'hono';
 
 export interface RepositoryMain {
   validateSendMail(params: EntityMain): Promise<void>;
-  sendMail(c: Context, body: EntityMain): Promise<void>;
+  sendMail(c: Context, body: EntityMain, authUser: string): Promise<void>;
 }